Revised substantially and prepared for gradual release.
This branch was created by copying the code from the Dart main
repository into the fresh github repository, and finishing the
ongoing refactorization to use the adjusted dart:mirrors APIs.

This API adjustment was considered already long ago in ongoing
discussions about dart:mirrors, and it was deemed useful enough
for the main usage cases (such as serialization and polymer) to
try it out here; it is simply about returning values directly
instead of returning mirrors of those values, and it applies to
such operations as 'invoke' and 'getField' on an InstanceMirror.

Note that the Type related mirror operations still return mirrors
(TypeMirror or ClassMirror) rather than values of type Type, because
they are typically used reflectively rather than directly (e.g., to
iterate through all subclass mirrors for a given class mirror, or
for inspecting the declarations for a given library or class mirror).

The refactoring to use the near-dart:mirrors APIs was accompanied by
the introduction of a bridging layer, smoker.dart and
smoker_implementation.dart, which is intended to make all the tests
(in test/common.dart and test/reflector.dart) look similar to the
ones in smoke/test/common.dart where they originated.  These files
serve as a proof-of-concept that it is possible to preserve a similar
usage as that associated with smoke.dart.

The directory layout was also refactored, in order to allow for a gradual
introduction of the provided functionality along with small, focused test
cases, and the change amounts to moving all files to '../old/..', such
that the standard directories and files are the only files still available
outside '../old/..', but all the code (the directories lib, lib/src, and
test, and the files therein plus Makefile) is still available inside
'../old/..'.  In short, the package is now essentially empty, apart from
the files in '../old/..'.

All paths have been adjusted such that 'cd old; make run' still runs all
tests in common.dart and reflector.dart successfully, in both the
dynamic version (ultimately using dart:mirrors) and the static version
(using hand written classes implementing mirrors without any use of
dart:mirrors --- serving as examples of the output that the transformer
will generate).

From here, the existing functionality will be extracted in manageably
sized pieces from '../old/..' and introduced into a fresh version of
the package, and when all useful pieces have been extracted the 'old'
directories will be deleted.

R=floitsch@google.com

Review URL: https://codereview.chromium.org/864283004