Implements a simple per file transform.
Completed mirrors.dart (to ensure good portability, it makes no sense
to leave out 'class Comment' and small stuff like that).

Added default implementation of all mirror classes in mirrors_unimpl.dart,
intended to derive static mirror classes from:  Anything which is in the
interface but not generated will then be an inherited method (so it is
implemented, and the program is compilable) which will fail with an
UnimplementedError at runtime.  This is how static mirrors should behave
when asked to do something that has not been generated.

Currently the generated code is almost a fixed string, it just includes
comments at the end listing the classes which have Reflectable derived
metadata, i.e., the classes for which we must generate static mirror
classes.

Added the file lib/src/reflectable_template.txt.  The intention is that
it will be read into a string, all place-holders will be replaced by
Dart code (e.g., maps, lists, classes), and the resulting string will
be stored as a file, which is the one that the transformed user
libraries will import (rather than reflectable.dart).

R=floitsch@google.com, sigurdm@google.com

Review URL: https://codereview.chromium.org/922363002