Used advice from Rico to add post-build/pre-test hook.
Basis: outside .test_config and tools/add_symlinks.dart, all files are
identical to the ones 'git cl land'ed and reverted on March 4.

Runs tests locally, although that's not particularly predictive of the
outcome at package bots.

Unresolved issues:
  - Couldn't find any examples showing '.test_config' in use. Is that
actually a name known by tools (like '.status') or is it just an example
name that must be "registered" somewhere, in order to let the file
have any effect? Note that tools/tests.py does _not_ apply this hook
when executed as part of the 'bottest' script locally.
  - Not sure whether a file like '.test_config' can be reformatted
to fit in 80 columns.  Surely JSON does not allow for splitting
strings (as in "Hello, world!" == "Hello, " "world!")?  Does it need to
fit in 80 columns?
  - tools/add_symlinks.dart adds the missing symlink; does this work on
Windows?  If not, what do we normally do on Windows when symlinks are
needed?  Do we simply turn of testing on windows (for all transformed
programs, because they all rely on having the missing symlink)?

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/981203009