Finds the prefix of the import of reflectable.
Just to push this CL through the system, even though I couldn't find
any need to _use_ the prefix.  I've added a test that puts a prefix
on the import of reflectable.dart.  The user-written code in there
needs to use the prefix in order to denote classes like Reflectable
and ReflectCapability, but no generated code needs to be changed.

We may 'git cl land' this if we know that there will be cases where the
prefix is needed, or we may just leave it on a branch "just in case".
Or you may have a case immediately where some prefix requires
different generated code?

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1004353004