Changed this description---the previous one talked about analyzer.dart and setup_analyzer.dart, but that whole excursion was dropped. Instead we made the approach used so far work again, by cheating code_transformers into delivering a Resolver for an AggregateTransform.
With that, the original plan has been completed, in that the transformation is now global rather than per library. Moreover, the code is now in place for adding missing imports FROM each library containing a class wherein we will generate code that refers to static mirror classes TO libraries containing those static mirror classes (which was the main motivation for changing from Transform to AggregateTransform in the first place).

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1067283002