Now prints an info message for each entry point.
This will not fix the problem with missing transformation on Windows,
but it will provide more information about the situation on all
platforms than we have currently. These info messages would not
be unreasonable to have for normal transformer runs, so we will
introduce them as a standard behavior now.

Review URL: https://codereview.chromium.org/1088193002