Attempted fix on comparison of paths.
"Attempted fix" rather than "Fixed" because we think we cannot test locally.

Changed two locations where path comparison is likely to have platform
issues (that is, it might break on Windows) because one of the compared
paths is obtained from an AssetId (where paths are always normalized to
use the separator '/') and the other one comes from a diffeent source,
such as the `fullName` method on an `Element` or `Source` object from the
analyzer.

Also added an invocation of 'path.normalize', which will eliminate
superfluous parts of paths (so 'some/./path/../otherpath' becomes
'some/otherpath'), which would improve on the ability of the package
to tolerate differences among paths that are only "syntactic" and hence
best ignored.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1051093006