Now runs invoke_test.
This includes a new piece of transformation: In static instance mirror
classes, an `invoke` method is generated for invocation of all methods
taking zero arguments (other cases left as UnimplementedError for now).

Note that the "Salt" (where names contain a large number like 9238478278
in order to make name clashes with existing names unlikely and name clashes
between generated names impossible) is currently disabled, because it
breaks the checktransform_test.dart test.  However, that will not be the
case any more when we switch to the new version of that test where only
a handful of the simplest transformations are checked by textual
comparison.

Added a unit test for relativePath, too.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1078223003