Moved everything into new "reflectable" subdir.
This version does not pass any tests, because it needs an update to
annotated_steps.py such that it will run a different kind of test
procedure for this special case: There is a `tools/annotated_steps.py`
in this github repo, so it wants to control the process itself
(because of the non-standard dir layout with a package and its test
package together in one repo), so we delegate to that *.py file to
do it. But updating the global annotated_steps.py is a step taken
separately.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1136713005