Change reflectable to use Strings instead of symbols everywhere
Also refactors transformer_implementation so that information is
gathered in a model before the transformation takes place. Hopefully
the result is a bit simpler.

The transformer_implementation creates a map of all the symbols of
named arguments for each reflector to be able to use Function.apply.

BUG=
R=eernst@google.com

Review URL: https://codereview.chromium.org/1149893008