Implement `.instanceMembers`.
Also implement support for invoke of operators.

The `owner` of an instance-member is not set.

The members are replicated in the declarations. We should perhaps make them
top-level fields, so they can be refenrenced from everywhere.

Applies on top of https://codereview.chromium.org/1181993003
BUG=
R=eernst@google.com

Review URL: https://codereview.chromium.org/1182083002.