Worked on support for fields.
Added a bunch of comments here and there, as I was digging into the
implementation and getting an understanding of the purposes/invariants.

Introduced VariableMirrorImpl, created them, used them in `declarations`.

Also adjusted reflectable/.status to make the comment correct, which is
probably also the intended semantics.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1212453004.