Implemented the single-list-of-mirrors approach.
The `memberMirrors` list is now a list of DeclarationMirror containing
both the VariableMirrors (at the front) and the MethodMirrors (as before),
and the indexing has been adjusted such that all method mirrors are
accessed at the required offset.

Added a bunch of comments along the way, in order to make the apparent
invariants explicit. In particular, I mentioned that `declarations` does
not mean the same thing everywhere, which might be a hint that we should
rename the non-mirror-semantics one in [ClassDomain].

Review URL: https://codereview.chromium.org/1225433005.