Implemented `invoker` on `ClassMirror`s.
Also corrected a number of imprecise parts of the dartdoc comments
about mirror operations, esp. in relation to the return types that
differ in reflectable/lib/mirrors.dart and in dart:mirrors.

Concept: An invoker is a closure which will accept a receiver and
return another closure which will call a method. So if you want to
call a dynamically selected method (the same one) on many different
objects, using `invoker` will provide such a closure once and for all,
whereas the "normal" approach (using `invoke` on instance mirrors) will
require a reflective operation for each receiver, and in practice often
multiple reflective operations for each invocation of said method.

Remaining issue: An invoker on an operator cannot be created in
pre-transformed code running on the VM (reported to github/sdk,
Area-VM as issue #23880, seems to be a bug that involves state
associated with operator names).

R=floitsch@google.com

Review URL: https://codereview.chromium.org/1249543004 .