Changed link_reflectable.dart to use native paths rather than Uri
Drilling down into the behavior of the Windows package bots (based
on lots of log files and on a reproduction so-far-as-possible on the
Windows laptop), it turns out that Uri.resolve produces a path ending in
  ...\reflectable\reflectable
where we should have had a path ending in
  ...\dart/third_party/pkg/reflectable/reflectable
(possibly with platform specific path separators, but this is a close
as I could get in the reproduction). Hence, I suspect that it is not
safe to use Uri's and relying on them to standardize on using POSIX
paths, and then using `toFilePath` at the end. The problem is probably
that URIs according to RFC 2396 do _not_ associate '/' with any particular
semantics (it does not differ from, say, '\\' or 'x').

As a consequence, I'm trying out a different basic approach: Use native
paths throughout, and then create them using the platform specific
separators as provided by the `path` package.

Will try it out now, `git cl land`, and revert if it makes the situation
worse (locally, there is no problem with the new code on Linux, and
the Windows bot is red since forever anyway, so it should be rather
benign as in terms of the bot color).

Review URL: https://codereview.chromium.org/1253913002 .