Changed projectRoot in link_reflectable.dart to use slash
Found a very small change that might solve the path manipulation
problem on Windows. As usual, it all runs fine on Linux.

After reading a large chunk of RFC 3986 about URIs and checking
various documentation (e.g., about class Uri), I concluded that
it should actually work to consider paths as URIs for a while, and
then get the actual (platform dependent) path using `toFilePath()`.
The inconsistency in `link_reflectable.dart` that (presumably) creates
the problem is the use of `args[0]` in `resolve`: On Windows, `args[0]`
will be a Windows path (relative, in the cases I've seen, but I cannot
rule out that an absolute path might be used in some context), but
it is used as a 'relative reference' (aka relative URI) in RFC 3986
lingo, and for that it must use slashes rather than backslashes to
separate its "path segments". If there is a drive letter in
`args[0]` and it is relative then it will fail, but I reckoned that this
is an unlikely corner case. If it is absolute then the Uri object will
look funny (e.g., "file:///C:/xxx/yyy/"), but it should be possible to
encode and decode it; if it is relative then everything looks "as one
would expect".

Will land and see if it helps, TBR.

Review URL: https://codereview.chromium.org/1254163003 .