Implemented upward-closure of set of reflectable classes.
Initiated because Jakob MacDonald hit a problem when trying to
port polymer to use reflectable.

The ability to use `superclass` on a class mirror (actually, a
`ClassMirrorImpl`) breaks down very frequently if there is no way
to create side-tags and no way to include superclasses among the
classes supported by a given reflector. For instance, `Object`
cannot be supported.

This CL implements support for adding extra classes to the set of
reflectable classes, such that it forms an upwards closed set
(if a class C is there and D is a superclass, it is also there).

This required a certain amount of refactoring, and many names have
been made private in order to make it easier to see the consequences
of various possible refactorings.

A tiny extra element: In response to another remark from Jakob,
the amount of noise created when an entry point is missing was
reduced (it's now an `info` rather than a `warning` message).

R=floitsch@google.com

Review URL: https://codereview.chromium.org/1257223006 .