Categorizes TODOs, added script to check this kind of categorization.
The `grep_todos` script will grep (skipping 'packages' and 'build') for
TODO, distinguishing between different values for the first "argument"
where I propose that we put the 'category' of the TODO. For instance,
a 'TODO(implement, ..)' indicates that something needs to be implemented,
e.g., because it is currently missing, incomplete, or buggy, and
'TODO(feature, ..)' indicates that something needs to be implemented, and
it is also a client-visible feature. It should be easy to maintain a
rather consistent approach, because `grep_todos` can find every TODO
which does not have a category among the ones that are mentioned in
reflectable/doc/todos.md. Another usage is a command like
`grep_todos '.*eernst'` which will find all TODOs for eernst (possibly
extended with a bit of noise).

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1266103007 .