Add a test for parameter-mirrors.
Still failing in transformed mode.

Fixes a bug in reflectable_implementation when the defaultValue is null.

The types of the parameters are not tested. The question
is how we want to represent these (returning a TypeMirror or the `Type` instance) what happens if the type is not itself reflectable? Is this ever used (it is only the static type, that cannot be trusted anyway).

BUG=
R=eernst@google.com

Review URL: https://codereview.chromium.org/1278493003 .