Clarified the status of implicit accessors; fixed several related bugs.
Added a test 'implicit_getter_setter_test.dart' which explores the
properties of implicit getters and setters for various kinds of
instance and static variable declarations.

Refactored the representation of mirrors in transformed code, in order
to enable unusual things for this purpose: The parameter of an implicit
setter has a special name, finding the owner of several of the relevant
mirrors used to involve some circularities (that we couldn't see how
to break with the old setup),

Deleted several declarations which are not used because of the
refactoring. Deleted code which used to create implicit accessor
mirrors (actually it was dead code, which was one of the bugs, but
now it's replaced by different code in a different location).

Implemented `staticMembers` on `ClassMirrorImpl` in order to support
static implicit accessors, and added various members earlier in the
pipeline to support it (`_staticMemberCache` and others).

Reclassified 'parameter_mirrors_test.dart' in .status: It is actually
blocked by the same issue #8 as some other tests (moving code for
the value of default arguments).

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1297503006 .