Bumping to version 0.1.2
This update just prepares the package for `pub publish`. There was a
version conflict that was created by an update to `code_transformers`
(which made the bots red even though we hadn't changed anything) and
we must publish the version of reflectable that doesn't trigger this
version conflict for others.

Will land it now, TBR.

Review URL: https://codereview.chromium.org/1300703005 .