Implements default values in mirrors.
Added implementation of methods to handle the operation where code
K is modified to K' such that K' has the same semantics in a library
L' as K had in L. It only covers some simple cases (literals, expressions
based on literals, e.g., `"4" + "2"`, and simple identifiers and
constructor invocations). This is, however, enough for the test cases
that we have so far (including 'new_instance_default_values_test.dart'
that we have kept in .status for a while). Hence, this mostly
resolves issue #8.

Generalized `returnType` to handle `dynamic`, just like `type` in variables
and parameters.

At this point, only two tests are disabled via .status.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1314473002 .