Modifies capability management to enforce `typeCapability`.
We used to allow a number of things related to types that are documented
in the design document to require a `typeCapability`, e.g., performing
`reflectType`. It is now enforced that these type related operations
will fail if there is no `typeCapability`. Moreover, several other kinds
of capabilities have been made subtypes of `TypeCapability`, such that
the inclusion of those capabilities will imply the inclusion of a
capability which `is TypeCapability`. For instance, there is no way to
use `declarations` unless we have `ClassMirror`s for some classes
(which is not the case if there is no `TypeCapability`), so users would
end up always having to write `declarationsCapability, typeCapability`.
With the new subtype relations this needless verbosity is avoided.

The changes do have an unexpected effect, though: The `dart2js` compiler
crashes during translation of build/test/mixin_test.dart (which does not
happen for the vm). This should be reported, of course, but right now
it is just indicated in '.status' that the crash is expected.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1340663002 .