Prevents generation of code where "... with ..." occurs as an identifier.
We have filtered out static members of mixin applications because they
do not exist semantically (and noone would be able to access them
without reflection anyway), but we were still generating tear-off
getters for them, and they would use the name of the mixin application
to access the member, and that's a string containing " with " (so the
resulting code is syntactically incorrect).

This CL cleans up the naming of mixin applications such that
transformed and untransformed code agree, and it avoids generating
those syntactically incorrect pieces of generated code.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1349793002 .