Enforces semantics of `typeRelationsCapability`.
There was an inconsistency in the treatment of mixins: The classes
delivering mixins (in `class C extends B with M`, `class M` delivers the
mixin) would have class mirrors (i.e., be supported) when they were
themselves annotated or matched a global quantifier etc., but also when
they were used in a mixin application which was reached by upwards
closure, if they were themselves under the bound, if any. In other
words, mixin-delivering classes were treated like superclasses.

This CL changes the treatment to match the design document better: A
`typeRelationsCapability` enables `mixin`, and hence the mirrors for
mixin-delivering classes are included when that capability is present,
whether or not the mixin-delivering classes are subclasses of a given
upper bound for an upwards closure.

This means that the use of `typeRelationsCapability` is required in
some cases where it was previously not required, but the overall
semantics is more consistent.

At the same time, the ability to exclude the upper bound in a
`SuperclassQuantifyCapability` was added, such that it is possible
to avoid including a very heavy class (such as `Element` in
'dart:html' and still get all its subclasses).

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1331423003 .