Fixes an invalid const value problem.
When generating import directives for the generated library, we used to
use the original name for the original entry point (that file is renamed
and the generated one gets its name); it should be replaced by the name
which that library is renamed to, and that is what we do now.

(Weirdly, that was not the actual problem, because a twisted path of
imports and exports meant that the wrong "import 'my_main.dart';" had
the same effect as the corrected
"import 'my_main_original_reflectable_main.dart';", but it is still
nicer to have the correct import than it is to rely on this twisted
coincidence).

Next, this CL introduces an extra step into `_extractConstantCode`,
such that it inlines the initializer of each private identifier it
encounters. (That turned out to be the actual problem that caused
the invalid constant values.)

A test has been added, using a constructor argument default value
(which is the case where we will move this kind of expression)
to test moving both a public and a private `const` identifier.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1356883002 .