Adds type annotation quantifier, remodels `_supportedClasses`.
With this CL, by default, the inclusion of a class mirror for a class
`C` does not imply that classes used in type annotations in `C` are
included (as they used to be when we had `declarationsCapability`).
This means that the set of supported classes can be much, much more
"frugal", but in return the client programmer must add some extra
code to include those extra classes that used to be supported: So
this is a BREAKING change. Here are the added features than enable
requesting these extra classes:

This CL adds a new `TypeAnnotationCapability({bool transitive: false})`
whose semantics is that classes used as type annotations are included,
one level deep if `transitive` is false and recursively if it is true.

This CL adds two shorthands for it: `typeAnnotationQuantifyCapability`
and `typeAnnotationDeepQuantifyCapability`.

Finally, this CL extensively remodels the code for computing
`_supportedClasses`: The algorithm has been refactored such that
large pieces of code solving one task have been moved to separate
functions/methods (depending on the need for `this`), and the
algorithm has been changed such that it computes closures as follows:

  - first it finds all directly supported classes
  - then it performs a subtype closure on that set
  - then it performs a superclass closure on that set
  - then it performs an annotation class extension on that set

where each 'then' step is performed if that feature has been
requested. In the last step 'extension' means a single iteration
adding annotation classes or an iteration which repeatedly adds
annotation classes, depending on the value of `transitive`.

All these changes are concerned with the pre-transformed code;
transformed code behaves as it did previously (which happens to
make all tests succeed). Next CL will include negative tests, such
that the distinction is detected.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1370463002 .