Splits `reflectableSupportsInstanceInvoke` and adds tests.
We used to use `reflectableSupportsInstanceInvoke` in order to check
not only whether a given member name could be used for invocation with
a given reflector and class, but also for filtering declarations to
report with `declarations`.

A 'TODO' indicated that this was a known issue to clarify, but now it
turned out to also cause buggy behavior in practice. This CL splits the
function such that we now also have `reflectableSupportsDeclaration`
which is used for `declarations` and works the same as
`reflectableSupportsInstanceInvoke` used to, and then changed
`reflectableSupportsInstanceInvoke` to scan up through
superclasses.

Is intended to be applied after inclusion of the pull request
https://github.com/dart-lang/reflectable/pull/41, such that the
solution given there gets adjusted a bit.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1381563011 .