Support for `reflectedType`, `reflectedReturnType`.
This CL adds support for `reflectedType` in variable mirrors and
parameter mirrors, and `reflectedReturnType` in method mirrors,
if enabled with `reflectedTypeCapability`. The idea is that it enables
users to get a small amount of support for reflection on these type
annotations, without paying the space penalty of generating all the
corresponding class mirrors.

A number of bugs were squashed during this process: `MixinApplication`
used to return `mixin.type` as its `type`; it is not easy to create a
suitable type, and `mixin.type` is clearly not the right one, so `type`
is now back to `_unimplemented()`, and some extra locations have
checks for `is MixinApplication`. `_typeCode` was split in two in
order to handle `MixinApplication` correctly, and enable a simpler
series of checks for cases where a `DartType` is available. The type
`void` will no longer be returned as a `FakeType`, and there is a
`can..` method to indicate that it will throw an exception if an
attempt is made to get such a type. `FakeType` use the qualified
name now, such that they won't incorrectly report equality for two
different fake types with the same name but from different
libraries.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1392903003 .