Creates `reflectors`, as a meta-meta feature enabling dynamic selection of the "mirror system".
When a library is expected to be used in a context where some useful
reflectors have already been created, it is difficult for that library
to find the reflectors, but it is very easy to return the set of all
reflectors from a static method in `Reflectable`. As a result, such a
library can now offer services using reflectable based reflection,
without requiring an `init` method to learn about which reflectors
there are, and without creating its own reflector (which may cause a
lot of information to be duplicated).

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1391543003 .