Using input from Stephen to improve on size of generated code.
Changed many literal lists from plain to `const` lists. Changed the
treatment of metadata to generate a `const` list in the first place,
rather than generating a non-const list and then passing it through
`UnmodifiableListView`.

Did not follow the advice to remove all type arguments on the list
literals.

All in all, the reduction in the output size is less than 1%, but the
code contains a number of extra `const` entities, so identical values
will be shared (which may matter much more in larger programs) and the
avoidance of a large number of `UnmodifiableListView` would also be
helpful in terms of startup speed.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1399053003 .