Followup on earlier review: Cleans up treatment of `getterMetadata`.
Introduces `MetadataEvaluator` and `_getMetadata{Class,Library}Factory`
such that the computation of metadata for the corresponding getter is
delayed, in cases where a `correspondingSetterQuantifyCapability`
allows access via such a getter. Uses this and a few other new
functions to avoid a handful of redundant pieces of code, removing a
couple of bugs along the way.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1410883002 .