Added dependency override to avoid looping analyzer
Currently, all package bots for reflectable are purple, which is caused
by a non-terminating `pub build --mode=debug test` process. It turns
out that we can avoid the issue by avoiding the most recent version of
the analyzer (taking instead version 0.26.1+14 which was used
previously). The problem arises when using code from 'code_transformers'
but it helps to use the old 'analyzer', so presumably the problem is in
the analyzer or at least in the interplay between the analyzer and
'code_transformer'. This CL will make it possible to continue to use
reflectable, and then we can find a real solution later, no matter
whether it means updates to 'analyzer', to 'code_transformers', to
'reflectable', or to several of the above.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1420593007 .