Fixes several problems with top-level declarations.
Introduces `..Attributes` and related constants for top level entities
(used to use `method` also for toplevel functions, which breaks `owner`
because it is unknown in which data structure that owner must be looked
up). Corrects `invokeSetter` to also accept a name without the '='
at top level, like it has always done for class members. Corrects
`MethodMirror.owner` to detect lack of capability and also to work for
top level entities. Corrects return type of `MethodMirror.owner` such
that it works for top level entities as well. Fixes `owner` in
`VariableMirror` in the same ways. Corrects treatment of libraries in
`ReflectorDomain` to work with top level entities. Corrects treatment
of class indices such that unsupported classes are handled gracefully.
Generalizes generation of code for `MethodMirrorImpl` such that it
works for top level entities (esp. top level variables and their
accessors needed changes, and `owner` related information needed to be
passed around and used). Corrects a large handful of types because
they were too "optimistic", esp. using `List` when only an `Iterable`
could actually be guaranteed (based on running the transformer in
checked mode). Corrects a number of locations based on failing
assertions (also due to the checked mode execution). Finally, corrects
the treatment of metadata such that also invocations of top level
getters (of a const value) can be evaluated at compile time, such
that `@c` will be recognized just as well as `@C()` when enabling
support for members and top level entities based on metadata.

Started out as a fix for issue #44, but no issues specifically associated
with >1 declared name with a single type annotation were uncovered.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1421713004 .