Starting from `isEnum`, adds support for several class properties.
This CL is intended to be used as a followup to the pull request about
`isEnum`: it adds support for that method in translated code.

Class properties in general (`isEnum`, `isPrivate`,
`isOriginalDeclaration`, etc.) turned out to be a very sparsely
implemented area. This CL creates a test 'class_property_test.dart'
which tests these properties, and implements support for it.

In particular, it changes many things in order to provide basic support
generic classes. Mirrors for non-generic classes are now able to report
that they do not have any type arguments/parameters, whether they
are the original declaration or not, and they can deliver the original
declaration. Missing (easy, probably) is the ability to report on
a non-empty list of type parameters and (much harder) the ability
to report on the actual type arguments in a parameterized type.
The latter may require som extra primitives.

The CL also adds checks to ensure that we do not ask for
the `reflectedType` on a mirror that doesn't have it.

`ClassMirrorImpl` has been made abstract, and it now has three
subclasses: `NonGenericClassMirrorImpl`, `GenericClass..`, and
`InstantiatedGenericClass..`; `Non..` is the simple case that we
have handled until now, `Generic..` is the generic class that
returns true on `originalDeclaration` and has no reflected type
(which we haven't done correctly so far), and `Instantiated..` is
the instantiated generic class ("parameterized type") which we
have not had so far, either. These three classes simplify the
situation around type arg/parm a lot, and only the harder
problems remain as TODOs.

`InstantiatedGenericClassMirrorImpl` could have been implemented
with a lot of forwarding (redundant code), or it could inherit
the code and preserve the structure of the data; I chose the
latter, but the trade-off is not obvious.

Code returning a type annotation will now create and return an
`Instantiated..` mirror; it does _not_ have information about
the actual type arguments, but it serves as a reminder about
which locations in the code we have to revise in order to obtain
true support for type arguments. Locations calling
`_createInstantiatedGenericClass` will then automatically be
required to provide that information, too. So we're "positioned"
well in order to improve on this.

Adjusted the treatment of library prefixes in generated code:
When imported names include names from 'dart:core' then we used
to generate `dart.core.SomeType`, which does not work. Replaced
that prefix by the empty string, and adjusted the treatment of
the '.' separator.

Finally, it turned out that `unit.types` does not include enum
classes, they're provided by `unit.enums`; taking that into
account removed a bug that we used to ignore enum classes.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1433973002 .